Basic movement and combat implementation
The enemy start chasing the player when the player gets too close. The enemy stops chasing if the player runs outside of the enemy chasing range at which point the enemy returns to it's original spawn position. The enemy also moves at it's own speed.

Combat implementation where the enemy can hit us or we can hit the enemy and if the hit points go to 0 we destroy it and we get some experience. Still need to implement the player death.

Also some pushback has been added when enemy or player collide.