Key and door mechanic
Addressed duplicate objects being made when switching scenes (player character and floating text).

Added key and door mechanic to transition between scenes. Made the door using 4 sprites that switch from closed to open sprites. Each door also specifies which key opens it so it will only open when the player comes into contact with the door while carrying the correct key. The door sprites can be changed in the inspector to easily have a working door in a different style.