npc text, healing fountain, health bar, destructable crates
Added a healing fountain that heals the player when they are close to it. Added some simple destructable crates.
Added a healthbar to the main screen UI so the player doesn't need to go into the status screen.
Added a script so NPCs, monsters or other objects can say can notify the player when they come in contact with them.