Tutorial level
Made a basic tutorial level with most of the core functionality.

The player is spawned and can interact with a door and an npc. The door tells the player they need a key. The npc tells the player to kill a skeleton that spawns. The key doesn't drop from the skeleton. But when the skeleton dies the player can destroy 2 crates which drop the key.